backend/requirements.txt
aioredis==1.1.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
certifi==2018.8.24
cffi==1.11.5
channels==2.1.3
channels-redis==2.3.0
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
cryptography==2.3.1
daphne==2.2.2
Django==2.1.2
django-cors-headers==2.4.0
django-debug-toolbar==1.10.1
djangorestframework==3.8.2
h2==3.0.1
hiredis==0.2.0
hpack==3.0.0
httpie==0.9.9
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
msgpack==0.5.6
priority==1.3.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
Pygments==2.2.0
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pypiwin32==223
pytz==2018.5
pywin32==224
requests==2.20.1
service-identity==17.0.0
six==1.11.0
sqlparse==0.2.4
Twisted==18.7.0
txaio==18.8.1
urllib3==1.23
zope.interface==4.5.0
